requirements.txt
boto3
numpy
jinja2
pyyaml
defusedcsv
